+++ src/main/java/me/coderfrish/ITag.java
@@ -1,4 +1,4 @@
-package io.github.xiefrish2021;
+package me.coderfrish;
 
 public interface ITag {
     default TagType type() {