+++ src/main/java/me/coderfrish/NBT.java
@@ -1,14 +1,17 @@
-package io.github.xiefrish2021;
+package me.coderfrish;
 
-import io.github.xiefrish2021.core.NBTException;
-import io.github.xiefrish2021.core.ObjectNBTMapper;
-import io.github.xiefrish2021.core.SNBTReader;
-import io.github.xiefrish2021.tag.compound.CompoundTag;
-import io.github.xiefrish2021.util.ReaderUtil;
-import io.github.xiefrish2021.util.WriteUtil;
-import org.jetbrains.annotations.ApiStatus;
+import me.coderfrish.core.NBTException;
+import me.coderfrish.core.ObjectNBTMapper;
+import me.coderfrish.core.SNBTReader;
+import me.coderfrish.tag.compound.CompoundTag;
+import me.coderfrish.util.ReaderUtil;
+import me.coderfrish.util.WriteUtil;
 
 import java.io.*;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
  * @author Frish2021
@@ -21,7 +24,7 @@ public final class NBT {
      * @param object Java Bean.
      * The authors of this method cannot guarantee its stability and may be removed in the future.
      */
-    @ApiStatus.Experimental
+    @Experimental
     public static void writeNamedNBT(String name, Object object, OutputStream out) {
         writeNamedNBT(name, new ObjectNBTMapper(object).toNBT(), out);
     }
@@ -31,7 +34,7 @@ public final class NBT {
      * @param object Java Bean.
      * The authors of this method cannot guarantee its stability and may be removed in the future.
      */
-    @ApiStatus.Experimental
+    @Experimental
     public static void writeUnnamedNBT(Object object, OutputStream out) {
         writeUnnamedNBT(new ObjectNBTMapper(object).toNBT(), out);
     }
@@ -96,4 +99,8 @@ public final class NBT {
             throw new NBTException(e);
         }
     }
+
+    @Retention(RetentionPolicy.CLASS)
+    @Target(ElementType.METHOD)
+    private @interface Experimental {}
 }