+++ src/main/java/me/coderfrish/TagType.java
@@ -1,4 +1,4 @@
-package io.github.xiefrish2021;
+package me.coderfrish;
 
 public enum TagType {
     END,